Fix webhook failing to remove entries
- The metadata-webhook was failing to remove entries. It determines
the subject to remove by analyzing the deleted files path. The logic
merely stripped the ".json" suffix from the path, but failed to
consider leading directories. As a result, if the file
"mappings/abcdef.json" was removed, it would try to remove
"mappings/abcdef" from the database, which would not work. The logic
now correctly takes the base name of the file, i.e.:

  takeBaseName "/directory/abcdef.ext" == "abcdef"