Use file-based store instead of Postgres DB
- Store metadata in the file system instead of a Postgres DB. PR #24
demonstrates that a large user of memory is the Postgres store,
specifically building the query for large batch-style requrests. This
could be investigated further, but a quick fix is simply to use a
file-based store. There is no need to build a query, and files can be
looked up quickly using their filename.